Armed units now have a target and can "attack" and getDistance now returns int instead of double

+++ BWAPI4J/src/main/java/org/openbw/bwapi4j/Position.java
@@ -34,7 +34,7 @@ public class Position {
      * Returns the distance as BW would.
      * This is ported from BWAPI's getApproxDistance method.
      */
-    public double getDistance(final Position position) {
+    public int getDistance(final Position position) {
     	int min = abs(getX() - position.getX());
     	int max = abs(getY() - position.getY());
